Database/requirements.txt
aiohttp==3.8.3
fastapi==0.95.1
filterpy==1.4.5
jose==1.0.0
matplotlib==3.7.1
motor==3.1.2
numpy==1.24.3
omegaconf==2.2.2
Pillow==9.5.0
psutil==5.8.0
pydantic==1.10.7
pymongo==4.3.3
python_jose==3.3.0
rasterio==1.3.6
scipy==1.10.1
shapely==2.0.1
starlette==0.26.1
uvicorn==0.22.0
hydra-core==1.2.0
python-multipart==0.0.6
websockets==11.0.1
gunicorn==20.1.0
Jinja2==3.1.
requests==2.26.0

